aura_backend/requirements.txt
fastapi>=0.110,<0.120
uvicorn[standard]>=0.23,<0.32
langchain>=0.2.0,<0.3.0
langchain-openai>=0.1.7,<0.2.0
pydantic>=2.0,<3.0
python-dotenv>=1.0,<2.0
